alloy_trie
Module nodes

Various branch nodes produced by the hash builder.

Structs

BranchNode

A branch node in an Ethereum Merkle Patricia Trie.

BranchNodeCompact

A struct representing a branch node in an Ethereum trie.

BranchNodeRef

A reference to BranchNode and its state mask. NOTE: The stack may contain more items that specified in the state mask.

ExtensionNode

An extension node in an Ethereum Merkle Patricia Trie.

ExtensionNodeRef

Reference to the extension node. See ExtensionNode from more information.

LeafNode

A leaf node represents the endpoint or terminal node in the trie. In other words, a leaf node is where actual values are stored.

LeafNodeRef

Reference to the leaf node. See LeafNode from more information.

RlpNode

An RLP-encoded node.

Enums

TrieNode

Enum representing an MPT trie node.

Constants

CHILD_INDEX_RANGE

The range of valid child indexes.

Functions

encode_path_leaf

Encodes a given path leaf as a compact array of bytes.

rlp_nodeDeprecated

Given an RLP-encoded node, returns it either as rlp(node) or rlp(keccak(rlp(node))).

word_rlpDeprecated

Optimization for quick RLP-encoding of a 32-byte word.