alloy_trie

Module proof

Proof verification logic.

Structs

DecodedProofNodes

A wrapper struct for trie node key to RLP encoded trie node.

DecodedProofRetainer

Proof retainer is used to store proofs during merkle trie construction. It is intended to be used within the HashBuilder.

ProofNodes

A wrapper struct for trie node key to RLP encoded trie node.

ProofRetainer

Proof retainer is used to store proofs during merkle trie construction. It is intended to be used within the HashBuilder.

Enums

ProofVerificationError

Error during proof verification.

Functions

verify_proof

Verify the proof for given key value pair against the provided state root.