alloy_trie

Module root

Functions

adjust_index_for_rlp

Adjust the index of an item for rlp encoding.

ordered_trie_root

Compute a trie root of the collection of rlp encodable items.

ordered_trie_root_with_encoder

Compute a trie root of the collection of items with a custom encoder.

state_root

Ethereum specific trie root functions. Calculates the root hash of the state represented as MPT.

state_root_ref_unhashed

Ethereum specific trie root functions. Hashes and sorts account keys, then proceeds to calculating the root hash of the state represented as MPT. See state_root_unsorted for more info.

state_root_unhashed

Ethereum specific trie root functions. Hashes and sorts account keys, then proceeds to calculating the root hash of the state represented as MPT. See state_root_unsorted for more info.

state_root_unsorted

Ethereum specific trie root functions. Sorts the hashed account keys and calculates the root hash of the state represented as MPT. See state_root for more info.

storage_root

Ethereum specific trie root functions. Calculates the root hash of account storage trie.

storage_root_unhashed

Ethereum specific trie root functions. Hashes storage keys, sorts them and them calculates the root hash of the storage trie. See storage_root_unsorted for more info.

storage_root_unsorted

Ethereum specific trie root functions. Sorts and calculates the root hash of account storage trie. See storage_root for more info.