Struct alloy_trie::TrieMask

pub struct TrieMask(/* private fields */);

A struct representing a mask of 16 bits, used for Ethereum trie operations.

Masks in a trie are used to efficiently represent and manage information about the presence or absence of certain elements, such as child nodes, within a trie. Masks are usually implemented as bit vectors, where each bit represents the presence (1) or absence (0) of a corresponding element.

Implementations

impl TrieMask

pub const fn new(inner: u16) -> Self

Creates a new TrieMask from the given inner value.

pub const fn get(self) -> u16

Returns the inner value of the TrieMask.

pub const fn from_nibble(nibble: u8) -> Self

Creates a new TrieMask from the given nibble.

pub fn is_subset_of(self, other: Self) -> bool

Returns true if the current TrieMask is a subset of other.

pub const fn is_bit_set(self, index: u8) -> bool

Returns true if a given bit is set in a mask.

pub const fn is_empty(self) -> bool

Returns true if the mask is empty.

pub const fn count_bits(self) -> u8

Returns the number of bits set in the mask.

pub const fn first_set_bit_index(self) -> Option<u8>

Returns the index of the first bit set in the mask, or None if the mask is empty.

pub fn set_bit(&mut self, index: u8)

Set bit at a specified index.

pub fn unset_bit(&mut self, index: u8)

Unset bit at a specified index.